NO-ISSUE Move the job implementation to generator
Not sure why the interface was in a separate package
